# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Telegram Pulse API
  version: 1.0.0
  summary: Curated REST contract for UU integration on top of custom auth.
  description: |
    Multi-tenant API for Telegram analytics, ads intelligence, trackers, rankings,
    mini apps, account management, team management, billing, and exports.
servers:
  - url: https://api.telegrampulse.example.com
    description: Production
  - url: https://staging-api.telegrampulse.example.com
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /v1.0/users/me:
    get:
      tags: [Account]
      operationId: getUserMe
      summary: Get current user profile.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    patch:
      tags: [Account]
      operationId: updateUserMe
      summary: Update profile fields.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MeUpdateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1.0/users/me/preferences:
    get:
      tags: [Account]
      operationId: getUserMyPreferences
      summary: Get user preferences.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPreferencesEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    patch:
      tags: [Account]
      operationId: updateUserMyPreferences
      summary: Update user preferences.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserPreferencesUpdateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPreferencesEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1.0/users/me/notifications:
    get:
      tags: [Account]
      operationId: getUserMyNotifications
      summary: Get notification settings.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationSettingsEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    patch:
      tags: [Account]
      operationId: updateUserMyNotifications
      summary: Update notification settings.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NotificationSettingsUpdateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationSettingsEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /v1.0/accounts/{accountId}/subscription:
    get:
      tags: [Billing]
      operationId: getAccountSubscription
      summary: Get active subscription.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubscriptionEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    patch:
      tags: [Billing]
      operationId: updateAccountSubscription
      summary: Change subscription plan/options.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubscriptionUpdateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubscriptionEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1.0/accounts/{accountId}/usage:
    get:
      tags: [Billing]
      operationId: getAccountUsage
      summary: Get account usage counters.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountUsageEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1.0/accounts/{accountId}/payment-methods:
    get:
      tags: [Billing]
      operationId: listPaymentMethods
      summary: List payment methods.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentMethodListEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Billing]
      operationId: addPaymentMethod
      summary: Attach payment method token.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentMethodCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentMethodEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1.0/accounts/{accountId}/invoices:
    get:
      tags: [Billing]
      operationId: listInvoices
      summary: List account invoices.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvoiceListEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1.0/accounts/{accountId}/invoices/{invoiceId}/download-url:
    get:
      tags: [Billing]
      operationId: getInvoiceDownloadUrl
      summary: Generate signed URL for invoice PDF.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
        - $ref: '#/components/parameters/InvoiceIdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvoiceDownloadEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AccountHeader:
      name: X-Account-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Account context. Must match path accountId when both are present.
    AccountPath:
      name: accountId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    InvoiceIdPath:
      name: invoiceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
    CursorParam:
      name: cursor
      in: query
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            error:
              code: unauthorized
              message: Missing or invalid bearer token.
              details: []
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            error:
              code: forbidden
              message: You do not have access to this account resource.
              details: []
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            error:
              code: not_found
              message: Resource was not found.
              details: []
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            error:
              code: validation_error
              message: Request validation failed.
              details:
                - field: er_min
                  issue: er_min must be less than or equal to er_max
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: true
    Page:
      type: object
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
    Meta:
      type: object
      additionalProperties: true
...
    MeProfile:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        first_name: { type: [string, 'null'] }
        last_name: { type: [string, 'null'] }
        full_name: { type: [string, 'null'] }
        email: { type: [string, 'null'], format: email }
        telegram_username: { type: [string, 'null'] }
        avatar_url: { type: [string, 'null'] }
    MeUpdateRequest:
      type: object
      properties:
        first_name: { type: string }
        last_name: { type: string }
        full_name: { type: string }
        telegram_username: { type: string }
        avatar_url: { type: string }
...
    UserPreferences:
      type: object
      properties:
        language_code: { type: string }
        timezone: { type: string }
        theme: { type: string, enum: [light, dark, system] }
    UserPreferencesUpdateRequest:
      type: object
      properties:
        language_code: { type: string }
        timezone: { type: string }
        theme: { type: string, enum: [light, dark, system] }
    NotificationSettings:
      type: object
      properties:
        email_notifications: { type: boolean }
        telegram_bot_alerts: { type: boolean }
        weekly_reports: { type: boolean }
        marketing_updates: { type: boolean }
        push_notifications: { type: boolean }
    NotificationSettingsUpdateRequest:
      type: object
      properties:
        email_notifications: { type: boolean }
        telegram_bot_alerts: { type: boolean }
        weekly_reports: { type: boolean }
        marketing_updates: { type: boolean }
        push_notifications: { type: boolean }
...
    Subscription:
      type: object
      properties:
        subscription_id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        plan_code: { type: string }
        status: { type: string }
        billing_state: { type: string }
        current_period_start: { type: [string, 'null'], format: date-time }
        current_period_end: { type: [string, 'null'], format: date-time }
        cancel_at_period_end: { type: boolean }
    SubscriptionUpdateRequest:
      type: object
      properties:
        plan_code: { type: string }
        cancel_at_period_end: { type: boolean }
    AccountUsage:
      type: object
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        channel_searches: { type: integer }
        event_trackers_count: { type: integer }
        api_requests_count: { type: integer }
        exports_count: { type: integer }
    PaymentMethod:
      type: object
      properties:
        payment_method_id: { type: string, format: uuid }
        brand: { type: [string, 'null'] }
        last4: { type: [string, 'null'] }
        exp_month: { type: [integer, 'null'] }
        exp_year: { type: [integer, 'null'] }
        is_default: { type: boolean }
        status: { type: string }
    PaymentMethodCreateRequest:
      type: object
      required: [provider_payment_method_token]
      properties:
        provider_payment_method_token: { type: string }
        make_default: { type: boolean, default: true }
    Invoice:
      type: object
      properties:
        invoice_id: { type: string, format: uuid }
        invoice_number: { type: [string, 'null'] }
        status: { type: string }
        currency: { type: string }
        amount_total: { type: number }
        period_start: { type: [string, 'null'], format: date }
        period_end: { type: [string, 'null'], format: date }
        issued_at: { type: [string, 'null'], format: date-time }
        paid_at: { type: [string, 'null'], format: date-time }
    InvoiceDownload:
      type: object
      properties:
        url: { type: string, format: uri }
        expires_at: { type: string, format: date-time }
...
    MeEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/MeProfile' }
        meta: { $ref: '#/components/schemas/Meta' }
    UserPreferencesEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/UserPreferences' }
        meta: { $ref: '#/components/schemas/Meta' }
    NotificationSettingsEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/NotificationSettings' }
        meta: { $ref: '#/components/schemas/Meta' }
...
    SubscriptionEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/Subscription' }
        meta: { $ref: '#/components/schemas/Meta' }
    AccountUsageEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/AccountUsage' }
        meta: { $ref: '#/components/schemas/Meta' }
    PaymentMethodEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/PaymentMethod' }
        meta: { $ref: '#/components/schemas/Meta' }
    PaymentMethodListEnvelope:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/PaymentMethod' }
        meta: { $ref: '#/components/schemas/Meta' }
    InvoiceListEnvelope:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Invoice' }
        page: { $ref: '#/components/schemas/Page' }
        meta: { $ref: '#/components/schemas/Meta' }
    InvoiceDownloadEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/InvoiceDownload' }
        meta: { $ref: '#/components/schemas/Meta' }
...